requirements.txt
openai
python-dotenv
semantic-kernel
vt-py
python-telegram-bot
iso639-lang
